Update View.php
refs #13329

Might not fix #13329 but should improve performance when rendering multiple views per pageload. It can be eg many when viewing visitor log etc.

As the issue is referring to the "Segmented Visitor Log", it could actually fix that issue. Maybe for some reason the file cannot be written and it always re-generates it. 

`piwikVersionBasedCacheBuster ` is already cached in the method itself.